+++ play-dl/YouTube/index.ts
@@ -1,3 +1,4 @@
 export { stream, stream_from_info, YouTubeStream } from './stream';
 export * from './utils';
 export { YouTube } from './search';
+export { cookieHeaders } from './utils/cookie'
